added most of the functionality

+++ src/main/java/com/vyara/fantasy/data/entities/secondary/Rating.java
@@ -21,7 +21,7 @@ import javax.persistence.Table;
 public class Rating extends BaseEntity {
 
     @Column
-    private byte rating;
+    private Integer rating;
 
     @ManyToOne
     private Book book;
